refactor(ledger): remove leios as an era
